+++ code/libs/ajudante.py
@@ -8,14 +8,14 @@ from libs.datilografo import Escritor
 from libs.telepatia import Transmissor
 from libs.dado import Dado
 
-from libs.mpu import MPU
-from libs.barometro import Barometro
-from libs.gps import GPS
-from libs.pitot import Pitot
-from libs.arduino import Arduino
+from libs.sensors.mpu import MPU
+from libs.sensors.barometro import Barometro
+from libs.sensors.gps import GPS
+from libs.sensors.pitot import Pitot
+from libs.sensors.arduino import Arduino
 
-from libs.celula import Celula
-from libs.balanca import Balanca
+from libs.sensors.celula import Celula
+from libs.sensors.balanca import Balanca
 
 class Ajudante(object):
