+++ code/libs/configurador.py
